Use type "int" instead of "int64" for "store_id"

+++ collector/creditmemos_response.go
@@ -4,7 +4,7 @@ import "fmt"
 
 type creditmemosResponse struct {
 	Items []struct {
-		StoreId int64 `json:"store_id"`
+		StoreId int `json:"store_id"`
 		State CreditMemoState `json:"state"`
 	} `json:"items"`
 }
